Skiller ut Table, tegnes inni en Suspense med Await
